flow2text/pyproject.toml
[tool.poetry]
name = "flow2text"
version = "0.1.0"
description = ""
authors = ["Taylor Henderson"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
matplotlib = "^3.7.1"
pandas = "^2.0.1"
scikit-learn = "^1.2.2"
fast-pl-py = "^0.3.2"
fastdtw = "^0.3.4"
statsmodels = "^0.14.0"
tqdm = "^4.65.0"
pyyaml = "^6.0.1"
cython = "3.0.8"
ripser = "0.6.4"
pyarrow = "^15.0.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
